late night commit

+++ include/mymath.h
@@ -13,16 +13,16 @@
 
 ////////////////////////////////VEC2/////////////////////////////////////////
 struct vec2 {
-	float x, y;
-	explicit vec2(float _x = 0, float _y = 0):x(_x),y(_y){}
+	double x, y;
+	explicit vec2(double _x = 0, double _y = 0):x(_x),y(_y){}
 
-	vec2 operator*(float num) const { return vec2(x * num, y * num); }
-	vec2 operator/(float num) const { return vec2(x / num, y / num); }
+	vec2 operator*(double num) const { return vec2(x * num, y * num); }
+	vec2 operator/(double num) const { return vec2(x / num, y / num); }
 	vec2 operator-(const vec2& v) const { return vec2(x - v.x, y - v.y); }
 	vec2 operator+(const vec2& v) const { return vec2(x + v.x, y + v.y); }
 
-	inline void Rotate() { float a = y; y = -x; x = a; }
-	inline float length() const { return sqrt(x * x + y * y); }
+	inline void Rotate() { double a = y; y = -x; x = a; }
+	inline double length() const { return sqrt(x * x + y * y); }
 	inline void Normalize() {
 		vec2 newVec = vec2(x, y) / length();
 		x = newVec.x;
@@ -32,26 +32,26 @@ struct vec2 {
 	
 };
 
-inline float dot(const vec2& v1, const vec2& v2) { return (v1.x * v2.x + v1.y * v2.y); }
+inline double dot(const vec2& v1, const vec2& v2) { return (v1.x * v2.x + v1.y * v2.y); }
 
-inline float angle(const vec2& v1, const vec2& v2){return acos(dot(v1,v2)/(v1.length()*v2.length()));}
+inline double angle(const vec2& v1, const vec2& v2){return acos(dot(v1,v2)/(v1.length()*v2.length()));}
 
 ////////////////////////////////VEC3/////////////////////////////////////////
 
 struct vec3 {
 	
-	float x, y, z;
+	double x, y, z;
 
-	explicit vec3(float _x = 0, float _y = 0, float _z = 0) : x(_x), y(_y), z(_z) {}
+	explicit vec3(double _x = 0, double _y = 0, double _z = 0) : x(_x), y(_y), z(_z) {}
 	
 	
-	vec3 operator*(float num) const { return vec3(x * num, y * num, z * num); }
-	vec3 operator/(float num) const { return vec3(x / num, y / num, z / num); }
+	vec3 operator*(double num) const { return vec3(x * num, y * num, z * num); }
+	vec3 operator/(double num) const { return vec3(x / num, y / num, z / num); }
 	vec3 operator+(const vec3& v) const { return vec3(x + v.x, y + v.y, z + v.z); }
 	vec3 operator-(const vec3& v) const { return vec3(x - v.x, y - v.y, z - v.z); }
 	vec3 operator*(const vec3& v) const { return vec3(x * v.x, y * v.y, z * v.z); }
 
-	inline float length() const { return sqrtf(x*x+y*y+z*z); }
+	inline double length() const { return sqrtf(x*x+y*y+z*z); }
 
 	inline void normalize() { 
 		
@@ -69,7 +69,7 @@ inline std::ostream& operator<<(std::ostream& os, const vec3& v){
 
 inline vec3 normalize(const vec3& v){ return v / v.length();}
 
-inline float dot(const vec3& v1, const vec3& v2) { return (v1.x * v2.x + v1.y * v2.y + v1.z * v2.z); }
+inline double dot(const vec3& v1, const vec3& v2) { return (v1.x * v2.x + v1.y * v2.y + v1.z * v2.z); }
 
 inline vec3 cross(const vec3& v1, const vec3& v2) {
 	return vec3(v1.y * v2.z - v1.z * v2.y, v1.z * v2.x - v1.x * v2.z, v1.x * v2.y - v1.y * v2.x);
@@ -78,28 +78,28 @@ inline vec3 cross(const vec3& v1, const vec3& v2) {
 ////////////////////////////////VEC4/////////////////////////////////////////
 
 struct vec4 {
-	float x, y, z, w;
+	double x, y, z, w;
 
-	explicit vec4(float _x = 0, float _y = 0, float _z = 0, float _w = 0) : x(_x), y(_y), z(_z),w(_w) {}
-	explicit vec4(const vec3& v, float _w = 1) : x(v.x), y(v.y), z(v.z), w(_w) {}
+	explicit vec4(double _x = 0, double _y = 0, double _z = 0, double _w = 0) : x(_x), y(_y), z(_z),w(_w) {}
+	explicit vec4(const vec3& v, double _w = 1) : x(v.x), y(v.y), z(v.z), w(_w) {}
 
-	vec4 operator*(float num) const { return vec4(x * num, y * num, z * num, w * num); }
-	vec4 operator/(float num) const { return vec4(x / num, y / num, z / num, w / num); }
+	vec4 operator*(double num) const { return vec4(x * num, y * num, z * num, w * num); }
+	vec4 operator/(double num) const { return vec4(x / num, y / num, z / num, w / num); }
 	vec4 operator+(const vec4& v) const { return vec4(x + v.x, y + v.y, z + v.z, w + v.w); }
 	vec4 operator-(const vec4& v)  const { return vec4(x - v.x, y - v.y, z - v.z, w - v.w); }
 	vec4 operator*(const vec4& v) const { return vec4(x * v.x, y * v.y, z * v.z, w * v.w); }
 	void operator+=(const vec4 right) { x += right.x; y += right.y; z += right.z; w += right.w; }
 
-	float& operator[](int index) { return *(&x + index); }
-	float operator[](int index) const { return *(&x + index); }
+	double& operator[](int index) { return *(&x + index); }
+	double operator[](int index) const { return *(&x + index); }
 
 };
 
-inline float dot(const vec4& v1, const vec4& v2) {
+inline double dot(const vec4& v1, const vec4& v2) {
 	return (v1.x * v2.x + v1.y * v2.y + v1.z * v2.z + v1.w * v2.w);
 }
 
-inline vec4 operator*(float a, const vec4& v) {
+inline vec4 operator*(double a, const vec4& v) {
 	return vec4(v.x * a, v.y * a, v.z * a, v.w * a);
 }
 
@@ -117,10 +117,10 @@ public:
 		rows[2][0] = 0; rows[2][1] = 0; rows[2][2] = 1; rows[2][3] = 0;
 		rows[3][0] = 0; rows[3][1] = 0; rows[3][2] = 0; rows[3][3] = 1;
 	}
-	mat4(float m00, float m01, float m02, float m03,
-		float m10, float m11, float m12, float m13,
-		float m20, float m21, float m22, float m23,
-		float m30, float m31, float m32, float m33) {
+	mat4(double m00, double m01, double m02, double m03,
+		double m10, double m11, double m12, double m13,
+		double m20, double m21, double m22, double m23,
+		double m30, double m31, double m32, double m33) {
 		rows[0][0] = m00; rows[0][1] = m01; rows[0][2] = m02; rows[0][3] = m03;
 		rows[1][0] = m10; rows[1][1] = m11; rows[1][2] = m12; rows[1][3] = m13;
 		rows[2][0] = m20; rows[2][1] = m21; rows[2][2] = m22; rows[2][3] = m23;
@@ -150,7 +150,7 @@ public:
 
 	vec4& operator[](int i) { return rows[i]; }
 	vec4 operator[](int i) const { return rows[i]; }
-	operator float* () const { return (float*)this; }
+	operator double* () const { return (double*)this; }
 
 	mat4(vec4 it, vec4 jt, vec4 kt, vec4 ot) {
 		rows[0] = it; rows[1] = jt; rows[2] = kt; rows[3] = ot;
@@ -190,8 +190,8 @@ inline mat4 ScaleMatrix(vec3 v)
 	return m;
 }
 
-inline mat4 RotationMatrix(float angle, vec3 w) {
-	float c = cosf(angle), s = sinf(angle);
+inline mat4 RotationMatrix(double angle, vec3 w) {
+	double c = cosf(angle), s = sinf(angle);
 	w.normalize();
 	mat4 m;
 
@@ -214,7 +214,7 @@ inline void TransformPoint(vec3& v, const mat4& m) {
 }
 
 
-inline bool FE(float a, float b){
+inline bool FE(double a, double b){
 	return fabs(a-b)<0.0001f;
 }
